Position call to action elements
The elements in the call to action section needed to be centered
horizontally and vertically through the implementation of flex-box.
A change was made to the html so that the button was in a container
such that it could be positioned centrally. I added some basic
background and border coloring to the containers to aid in the
positioning process.