Position and style footer
I needed to center the footer text as well as change the backgroun
color. I ran into a problem of white space at the bottom of the page.
I discovered the body element was set to display block. In order to fix
the footer to the bottom of the page I implemented flex-box on the body.